requirements.txt
# FastAPI framework for building asynchronous APIs
fastapi==0.110.0

# ASGI server used to run FastAPI apps with performance enhancements
uvicorn[standard]==0.27.1

# Data validation and parsing library used by FastAPI (v2+ has major updates)
pydantic==2.7.1

# Python SDK for interacting with AWS S3 and MinIO-compatible storage
boto3==1.34.84

# Toolkit for WSI (Whole Slide Image) reading and processing (by TIA Centre)
tiatoolbox==1.3.0

# Fundamental package for array computing and matrix operations
numpy==1.26.4

# Image processing library used for saving patches (PNG, JPEG, etc.)
pillow==10.3.0

# Python bindings for reading whole slide images using OpenSlide backend
openslide-python==1.3.1
